vm: fix some size_t/uint64_t issues (25)

+++ src/cshared/buffer.cpp
@@ -16,8 +16,8 @@ Buffer::Buffer(const Buffer &other)
     this->length = other.length;
     if (other.data)
     {
-        this->data = (byte *)malloc(this->length);
-        memcpy(this->data, other.data, this->length);
+        this->data = (byte *)malloc((size_t)this->length);
+        memcpy(this->data, other.data, (size_t)this->length);
     }
 }
